MEED-6663 Refactor the scheduled tasks for evm contract
