+++ gamification-evm-services/src/main/resources/db/changelog/gamification-evm-rdbms.db.changelog-1.0.0.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
-
   This file is part of the Meeds project (https://meeds.io/).
   Copyright (C) 2020 - 2024 Meeds Association contact@meeds.io
   This program is free software; you can redistribute it and/or
@@ -14,7 +13,6 @@
   You should have received a copy of the GNU Lesser General Public License
   along with this program; if not, write to the Free Software Foundation,
   Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
-
 -->
 <databaseChangeLog xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
                    xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
